+++ MoviesData/IMovieRepository.cs
@@ -14,7 +14,7 @@ namespace MoviesData
 
         IReadOnlyList<Movie> GenreMovies(string Genre);
 
-        int TotalSales(string moviename);
+        double TotalSales(string moviename);
 
         IReadOnlyList<(Movie, int, int, string)> ActorGenreMovies(string firstName, string lastName, string genre, int minScore, int maxScore);
     }
